requirements.txt
google-cloud-aiplatform
python-dotenv
urllib3==1.26.18